+++ alicloud/import_alicloud_router_interface_test.go
@@ -10,7 +10,10 @@ func TestAccAlicloudRouterInterface_import(t *testing.T) {
 	resourceName := "alicloud_router_interface.interface"
 
 	resource.Test(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheck(t) },
+		PreCheck: func() {
+			testAccPreCheck(t)
+			testAccPreCheckWithAccountSiteType(t, DomesticSite)
+		},
 		Providers:    testAccProviders,
 		CheckDestroy: testAccCheckVpcDestroy,
 		Steps: []resource.TestStep{