+++ alicloud/resource_alicloud_vpn_gateway_test.go
@@ -103,6 +103,7 @@ func TestAccAlicloudVpnGateway_basic(t *testing.T) {
 	resource.Test(t, resource.TestCase{
 		PreCheck: func() {
 			testAccPreCheck(t)
+			testAccPreCheckWithAccountSiteType(t, IntlSite)
 		},
 
 		// module name
@@ -139,6 +140,7 @@ func TestAccAlicloudVpnGateway_update(t *testing.T) {
 	resource.Test(t, resource.TestCase{
 		PreCheck: func() {
 			testAccPreCheck(t)
+			testAccPreCheckWithAccountSiteType(t, IntlSite)
 		},
 		Providers:    testAccProviders,
 		CheckDestroy: testAccCheckVpnGatewayDestroy,